+++ wezterm-gui/src/termwindow/mouseevent.rs
@@ -1,17 +1,22 @@
+use crate::termwindow::TermWindowNotif;
 use crate::tabbar::TabBarItem;
 use crate::termwindow::keyevent::window_mods_to_termwiz_mods;
-use crate::termwindow::{MouseCapture, PositionedSplit, ScrollHit, UIItem, UIItemType, TMB};
+use crate::termwindow::{
+    GuiWin, MouseCapture, PositionedSplit, ScrollHit, UIItem, UIItemType, TMB,
+};
 use ::window::{
     MouseButtons as WMB, MouseCursor, MouseEvent, MouseEventKind as WMEK, MousePress, WindowOps,
     WindowState,
 };
-use config::keyassignment::{MouseEventTrigger, SpawnTabDomain};
+use config::keyassignment::{KeyAssignment, MouseEventTrigger, SpawnTabDomain};
 use config::MouseEventAltScreen;
 use mux::pane::{Pane, WithPaneLines};
 use mux::tab::SplitDirection;
 use mux::Mux;
+use mux_lua::MuxPane;
 use std::convert::TryInto;
 use std::ops::Sub;
+use std::rc::Rc;
 use std::sync::Arc;
 use std::time::Duration;
 use termwiz::hyperlink::Hyperlink;
@@ -398,6 +403,55 @@ impl super::TermWindow {
         context.set_cursor(Some(MouseCursor::Arrow));
     }
 
+    fn do_new_tab_button_click(&mut self, button: MousePress) {
+        let pane = match self.get_active_pane_or_overlay() {
+            Some(pane) => pane,
+            None => return,
+        };
+        let action = match button {
+            MousePress::Left => Some(KeyAssignment::SpawnTab(SpawnTabDomain::CurrentPaneDomain)),
+            MousePress::Right => Some(KeyAssignment::ShowLauncher),
+            MousePress::Middle => None,
+        };
+
+        async fn dispatch_new_tab_button(
+            lua: Option<Rc<mlua::Lua>>,
+            window: GuiWin,
+            pane: MuxPane,
+            button: MousePress,
+            action: Option<KeyAssignment>,
+        ) -> anyhow::Result<()> {
+            let default_action = match lua {
+                Some(lua) => {
+                    let args = lua.pack_multi((
+                        window.clone(),
+                        pane,
+                        format!("{button:?}"),
+                        action.clone(),
+                    ))?;
+                    config::lua::emit_event(&lua, ("new-tab-button-click".to_string(), args))
+                        .await
+                        .map_err(|e| {
+                            log::error!("while processing new-tab-button-click event: {:#}", e);
+                            e
+                        })?
+                }
+                None => true,
+            };
+            if let (true, Some(assignment)) = (default_action, action) {
+                log::info!("do it {assignment:?}");
+                window.window.notify(TermWindowNotif::PerformAssignment{ pane_id: pane.0, assignment });
+            }
+            Ok(())
+        }
+        let window = GuiWin::new(self);
+        let pane = MuxPane(pane.pane_id());
+        promise::spawn::spawn(config::with_lua_config_on_main_thread(move |lua| {
+            dispatch_new_tab_button(lua, window, pane, button, action)
+        }))
+        .detach();
+    }
+
     pub fn mouse_event_tab_bar(
         &mut self,
         item: TabBarItem,
@@ -410,7 +464,7 @@ impl super::TermWindow {
                     self.activate_tab(tab_idx as isize).ok();
                 }
                 TabBarItem::NewTabButton { .. } => {
-                    self.spawn_tab(&SpawnTabDomain::CurrentPaneDomain);
+                    self.do_new_tab_button_click(MousePress::Left);
                 }
                 TabBarItem::None | TabBarItem::LeftStatus | TabBarItem::RightStatus => {
                     // Potentially starting a drag by the tab bar
@@ -427,17 +481,17 @@ impl super::TermWindow {
                 TabBarItem::Tab { tab_idx, .. } => {
                     self.close_specific_tab(tab_idx, true);
                 }
-                TabBarItem::NewTabButton { .. }
-                | TabBarItem::None
-                | TabBarItem::LeftStatus
-                | TabBarItem::RightStatus => {}
+                TabBarItem::NewTabButton { .. } => {
+                    self.do_new_tab_button_click(MousePress::Middle);
+                }
+                TabBarItem::None | TabBarItem::LeftStatus | TabBarItem::RightStatus => {}
             },
             WMEK::Press(MousePress::Right) => match item {
                 TabBarItem::Tab { .. } => {
                     self.show_tab_navigator();
                 }
                 TabBarItem::NewTabButton { .. } => {
-                    self.show_launcher();
+                    self.do_new_tab_button_click(MousePress::Right);
                 }
                 TabBarItem::None | TabBarItem::LeftStatus | TabBarItem::RightStatus => {}
             },